+++ Changelog.md
@@ -5,6 +5,7 @@
 - Provide contextual menu to transform routes. In the `New Camel file` menu transform the right clicked files or all the file in a clicked folder.
 - Update default Camel version used for Camel JBang from 4.8.0 to 4.8.1
 - Update default Camel Catalog version from 4.8.0 to 4.8.1
+- Update Camel Quarkus Catalog from 3.15.0 to 3.16.0
 
 ## 1.5.0