+++ Changelog.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## 1.7.0
+
 ## 1.6.0
 
 - Provide contextual menu to transform routes. In the `New Camel file` menu transform the right clicked files or all the file in a clicked folder.